+++ src/jiratui/tests/test_app.py
@@ -1,5 +1,6 @@
 from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
+from pydantic import SecretStr
 import pytest
 
 from jiratui.api_controller.controller import APIController, APIControllerResponse
@@ -8,6 +9,7 @@ from jiratui.config import ApplicationConfiguration
 from jiratui.models import JiraServerInfo
 from jiratui.widgets.config_info import ConfigFileScreen
 from jiratui.widgets.help import HelpScreen
+from jiratui.models import JiraServerInfo, WorkItemsSearchOrderBy
 from jiratui.widgets.quit import QuitScreen
 from jiratui.widgets.screens import MainScreen
 from jiratui.widgets.server_info import ServerInfoScreen
@@ -19,8 +21,10 @@ def app_with_unrecognized_config_theme() -> JiraApp:
     config_mock.configure_mock(
         jira_api_base_url='foo.bar',
         jira_api_username='foo',
-        jira_api_token='bar',
+        jira_api_token=SecretStr('bar'),
         jira_api_version=3,
+        use_bearer_authentication=False,
+        cloud=True,
         ignore_users_without_email=True,
         default_project_key_or_id=None,
         jira_account_id=None,
@@ -31,6 +35,8 @@ def app_with_unrecognized_config_theme() -> JiraApp:
         log_file='',
         log_level='WARNING',
         theme='foo',
+        ssl=None,
+        search_results_default_order=WorkItemsSearchOrderBy.CREATED_DESC,
     )
     app = JiraApp(config_mock)
     app.api = APIController(config_mock)
@@ -44,8 +50,10 @@ def app_with_input_and_config_theme() -> JiraApp:
     config_mock.configure_mock(
         jira_api_base_url='foo.bar',
         jira_api_username='foo',
-        jira_api_token='bar',
+        jira_api_token=SecretStr('bar'),
         jira_api_version=3,
+        use_bearer_authentication=False,
+        cloud=True,
         ignore_users_without_email=True,
         default_project_key_or_id=None,
         jira_account_id=None,
@@ -56,6 +64,8 @@ def app_with_input_and_config_theme() -> JiraApp:
         log_file='',
         log_level='WARNING',
         theme='flexoki',
+        ssl=None,
+        search_results_default_order=WorkItemsSearchOrderBy.CREATED_DESC,
     )
     app = JiraApp(config_mock, user_theme='monokai')
     app.api = APIController(config_mock)
@@ -69,8 +79,10 @@ def app_with_input_theme() -> JiraApp:
     config_mock.configure_mock(
         jira_api_base_url='foo.bar',
         jira_api_username='foo',
-        jira_api_token='bar',
+        jira_api_token=SecretStr('bar'),
         jira_api_version=3,
+        use_bearer_authentication=False,
+        cloud=True,
         ignore_users_without_email=True,
         default_project_key_or_id=None,
         jira_account_id=None,
@@ -81,6 +93,8 @@ def app_with_input_theme() -> JiraApp:
         log_file='',
         log_level='WARNING',
         theme=None,
+        ssl=None,
+        search_results_default_order=WorkItemsSearchOrderBy.CREATED_DESC,
     )
     app = JiraApp(config_mock, user_theme='monokai')
     app.api = APIController(config_mock)
@@ -94,8 +108,10 @@ def app_without_config_theme() -> JiraApp:
     config_mock.configure_mock(
         jira_api_base_url='foo.bar',
         jira_api_username='foo',
-        jira_api_token='bar',
+        jira_api_token=SecretStr('bar'),
         jira_api_version=3,
+        use_bearer_authentication=False,
+        cloud=True,
         ignore_users_without_email=True,
         default_project_key_or_id=None,
         jira_account_id=None,
@@ -106,6 +122,8 @@ def app_without_config_theme() -> JiraApp:
         log_file='',
         log_level='WARNING',
         theme=None,
+        ssl=None,
+        search_results_default_order=WorkItemsSearchOrderBy.CREATED_DESC,
     )
     app = JiraApp(config_mock)
     app.api = APIController(config_mock)
@@ -119,8 +137,10 @@ def app() -> JiraApp:
     config_mock.configure_mock(
         jira_api_base_url='foo.bar',
         jira_api_username='foo',
-        jira_api_token='bar',
+        jira_api_token=SecretStr('bar'),
         jira_api_version=3,
+        use_bearer_authentication=False,
+        cloud=True,
         ignore_users_without_email=True,
         default_project_key_or_id=None,
         jira_account_id=None,
@@ -131,6 +151,8 @@ def app() -> JiraApp:
         log_file='',
         log_level='WARNING',
         theme='dracula',
+        ssl=None,
+        search_results_default_order=WorkItemsSearchOrderBy.CREATED_DESC,
     )
     app = JiraApp(config_mock)
     app.api = APIController(config_mock)