+++ CHANGELOG.md
@@ -14,6 +14,12 @@ settings `search_results_page_filtering_enabled` and `search_results_page_filter
 - Use `flat` buttons for search and for confirming quitting the app.
 - Refactor the logic to navigate search result pages to use dynamic actions.
 - Add support for full-text search.
+- Add support for client-side certificates and SSL configuration
+- Add support for Bearer authentication
+- Add support for configuring bearer authentication via config variable `use_bearer_authentication`.
+- Add configuration variable `cloud` to distinguish between on-premises and cloud platform.
+- Add an option to define the default value of the `order by` widget.
+- Update README and docs with instructions on how to install the tool via Homebrew.
 
 ### Bug Fixes